Moved files around, added comments, added unit tests.
