+++ tools/parse/src/main/scala/org/allenai/nlpstack/parse/poly/core/Sentence.scala
@@ -1,7 +1,7 @@
 package org.allenai.nlpstack.parse.poly.core
 
 import org.allenai.common.immutable.Interval
-import org.allenai.nlpstack.parse.poly.fsm.{ Sculpture, MarbleBlock }
+import org.allenai.nlpstack.parse.poly.fsm.MarbleBlock
 import org.allenai.nlpstack.parse.poly.ml.FeatureVector
 
 import reming.DefaultJsonProtocol._
@@ -63,7 +63,7 @@ case class Sentence(tokens: IndexedSeq[Token]) extends MarbleBlock {
     // (bottom of stack) to the end of the sentence.
     var unmatchedParenIx = -1
     while (parenStack.length > 0) {
-      unmatchedParenIx = parenStack.pop
+      unmatchedParenIx = parenStack.pop()
     }
     if (unmatchedParenIx > -1) {
       parenIntervals = parenIntervals + Interval.closed(unmatchedParenIx, tokens.size - 1)
@@ -76,8 +76,15 @@ case class Sentence(tokens: IndexedSeq[Token]) extends MarbleBlock {
 object Sentence {
   implicit val sentenceJsonFormat = jsonFormat1(Sentence.apply)
 
+  /** Initializes a Sentence object from a string, by splitting on whitespace to create tokens.
+    *
+    * Note that the zeroth token of the initialized sentence will be a NexusToken.
+    *
+    * @param rawString the input string
+    * @return the corresponding Sentence object
+    */
   def initializeFromWhitespaceSeparatedString(rawString: String): Sentence = {
-    val tokens = rawString.split("\\s+") map { str => Token(Symbol(str)) }
+    val tokens = rawString.trim.split("\\s+") map { str => Token(Symbol(str)) }
     Sentence(NexusToken +: tokens.toIndexedSeq)
   }
 }
