+++ tools/parse/src/main/scala/org/allenai/nlpstack/parse/poly/core/Util.scala
@@ -46,6 +46,12 @@ object Util {
     JsonParser.read[T](new BufferedSource(uncompressedStream))
   }
 
+  /** Uses an NlpStack postagger to tag a Sentence object.
+    *
+    * @param sentence the Sentence to tag
+    * @param posTagger the nlpstack postagger to use
+    * @return a map from Sentence token indices to their POS tags
+    */
   def getPostaggedTokens(sentence: Sentence, posTagger: Postagger): Map[Int, PostaggedToken] = {
     val words: IndexedSeq[String] = sentence.tokens.tail map { tok => tok.word.name }
     val nlpStackTokens: IndexedSeq[NLPStackToken] =
