+++ tools/parse/src/main/scala/org/allenai/nlpstack/parse/poly/decisiontree/OmnibusTrainer.scala
@@ -3,8 +3,8 @@ package org.allenai.nlpstack.parse.poly.decisiontree
 class OmnibusTrainer()
     extends ProbabilisticClassifierTrainer {
 
-  val dtTrainer = new RandomForestTrainer(0, 12, 0.1f, EntropyGainMetric(0.0f), numThreads = 6)
-  val rfTrainer = new RandomForestTrainer(0, 12, 0.1f, EntropyGainMetric(0.0f), numThreads = 6)
+  val dtTrainer = new RandomForestTrainer(0, 12, 0.1f, MultinomialGainMetric(0.5f), numThreads = 6)
+  val rfTrainer = new RandomForestTrainer(0, 12, 0.1f, MultinomialGainMetric(0.5f), numThreads = 6)
 
   override def apply(data: FeatureVectorSource): ProbabilisticClassifier = {
     val trainer = data.classificationTask.filenameFriendlyName match {
