+++ tools/parse/src/main/scala/org/allenai/nlpstack/parse/poly/fsm/TransitionConstraint.scala
@@ -33,6 +33,10 @@ object TransitionConstraint {
   */
 trait ConstraintInterpretation extends ((State, StateTransition) => Boolean)
 
+/** The TrivialConstraintInterpretation returns false for any state/transition pair.
+  *
+  * This means that transitions are always considered applicable.
+  */
 class TrivialConstraintInterpretation extends ConstraintInterpretation {
   def apply(state: State, transition: StateTransition): Boolean = false
 }
