+++ tools/parse/src/main/scala/org/allenai/nlpstack/parse/poly/polyparser/TransitionParser.scala
@@ -2,13 +2,11 @@ package org.allenai.nlpstack.parse.poly.polyparser
 
 import org.allenai.common.Resource
 import org.allenai.nlpstack.parse.poly.core.{ Token, NexusToken, Sentence, Util }
-import org.allenai.nlpstack.parse.poly.fsm.{ ClassifierBasedCostFunctionFactory, TransitionConstraint }
-import reming.{ CompactPrinter, JsonParser, LazyFormat }
+import org.allenai.nlpstack.parse.poly.fsm.TransitionConstraint
+import reming.{ CompactPrinter, LazyFormat }
 import reming.DefaultJsonProtocol._
 
 import java.io.{ BufferedWriter, InputStream, File, FileWriter, PrintWriter }
-
-import scala.io.BufferedSource
 
 /** A TransitionParser implements a parsing algorithm for a transition-based parser. */
 abstract class TransitionParser {
